refactor: cleanup ef tests
