+++ exact-numbers/src/test/kotlin/xyz/lbres/exactnumbers/exactfraction/ExactFractionOperatorsTest.kt
@@ -3,7 +3,7 @@ package xyz.lbres.exactnumbers.exactfraction
 import kotlin.test.Test
 
 class ExactFractionOperatorsTest {
-    @Test fun testEfAdd() = runEfAddTests()
-    @Test fun testEfTimes() = runEfTimesTests()
-    @Test fun testEfPow() = runEfPowTests()
+    @Test fun testEfAdd() = runCommonPlusTests { ef1, ef2 -> efAdd(ef1, ef2) }
+    @Test fun testEfTimes() = runCommonTimesTests { ef1, ef2 -> efTimes(ef1, ef2) }
+    @Test fun testEfPow() = runCommonPowTests { ef1, ef2 -> efPow(ef1, ef2) }
 }
