+++ exact-numbers/src/test/kotlin/xyz/lbres/testutils/helpers.kt
@@ -31,11 +31,25 @@ fun assertExactFractionOverflow(type: String, value: ExactFraction, cast: () -> 
  * Assert that a test throws an Exception with the correct message
  *
  * @param message [String]: expected error message
- * @param test () -> Unit: the code to run
+ * @param test () -> Unit: test to run
  * @return [Exception] the exception that was thrown
  */
 inline fun <reified T : Exception> assertFailsWithMessage(message: String, test: () -> Unit): T {
     val error = assertFailsWith<T> { test() }
     assertEquals(message, error.message)
     return error
+}
+
+/**
+ * Validate that a test succeeds, and throw error if it fails
+ *
+ * @param errorMessage [String]: error message to throw if test fails
+ * @param test () -> Unit: test to run
+ */
+fun <T> assertSucceeds(errorMessage: String, test: () -> T) {
+    try {
+        test()
+    } catch (_: Exception) {
+        throw AssertionError(errorMessage)
+    }
 }
